Add abstract SqliteRepository
