+++ TokTok/Repositories/Mock/MockUserRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using TokTok.Models;
 
-namespace TokTok.Repositories
+namespace TokTok.Repositories.Mock
 {
     public class MockUserRepository : IUserRepository
     {
